refactor with utilities.js
