+++ TaskTok/static/js/apiRequests/functions.js
@@ -1,28 +1,6 @@
-function clearModal(){
-    var editModal = document.getElementById('editModal');
-    editModal.style.display = 'none';
-    var modalBody = document.getElementById('modal-body');
-    var modalFooter = document.getElementById('modal-footer')
-    modalBody.innerHTML = "";
-    modalFooter.innerHTML = "";
-}
+import { showToast, clearModal, getCookie } from './utilities.js';
 
-function getCookie(name) {
-    let cookieValue = null;
-    if (document.cookie && document.cookie !== '') {
-        const cookies = document.cookie.split(';');
-        for (let i = 0; i < cookies.length; i++) {
-            const cookie = cookies[i].trim();
-            if (cookie.substring(0, name.length + 1) === (name + '=')) {
-                cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
-                console.log(cookieValue);
-                break;
-            }
-        }
-    }
-    return cookieValue;
-  }
-  
+
   
 async function addTask() {
     const csrfAccessToken = getCookie('csrf_access_token');
@@ -457,17 +435,7 @@ async function listTask() {
         clearModal();
       }
   }
-  function showToast(message, duration = 3000) {
-    const toast = document.getElementById("toast-container");
-    const toastContent = document.getElementById("toast-user-content");
-    toastContent.textContent = message;
-    toast.classList.add("show");
 
-    // Hide the toast after 'duration' milliseconds
-    setTimeout(() => {
-        toast.classList.remove("show");
-    }, duration);
-}
   
   document.addEventListener('DOMContentLoaded', function(){
   //Use the DOMContentLoaded to make sure the DOM is fully loaded before trying to load our script.
