+++ TaskTok/static/js/utilities.js
@@ -1,0 +1,57 @@
+export function showToast(message, duration = 3000) {
+    const toast = document.getElementById("toast-container");
+    const toastContent = document.getElementById("toast-user-content");
+    toastContent.textContent = message;
+    toast.classList.add("show");
+
+    // Hide the toast after 'duration' milliseconds
+    setTimeout(() => {
+        toast.classList.remove("show");
+    }, duration);
+}
+
+export function clearModal(){
+    var editModal = document.getElementById('editModal');
+    editModal.style.display = 'none';
+    var modalBody = document.getElementById('modal-body');
+    var modalFooter = document.getElementById('modal-footer')
+    modalBody.innerHTML = "";
+    modalFooter.innerHTML = "";
+}
+
+export function getCookie(name) {
+    let cookieValue = null;
+    if (document.cookie && document.cookie !== '') {
+        const cookies = document.cookie.split(';');
+        for (let i = 0; i < cookies.length; i++) {
+            const cookie = cookies[i].trim();
+            if (cookie.substring(0, name.length + 1) === (name + '=')) {
+                cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
+                break;
+            }
+        }
+    }
+    return cookieValue;
+  }
+  
+  export function format_backend_datetime(dateString){
+    const date = new Date(dateString);
+
+    let day = date.getDate().toString().padStart(2, '0');
+    let month = (date.getMonth() + 1).toString().padStart(2, '0'); // getMonth() returns 0-11
+    let year = date.getFullYear();
+    let hours = date.getHours().toString().padStart(2, '0');
+    let minutes = date.getMinutes().toString().padStart(2, '0');
+
+    return `${month}/${day}/${year} ${hours}:${minutes}`;
+}
+
+export function enableVerticalScroll() {
+    const scrollContainer = document.querySelector('.archived-tasks-container');
+    if (scrollContainer) {
+        scrollContainer.addEventListener('wheel', (event) => {
+            event.preventDefault();
+            scrollContainer.scrollLeft += event.deltaY;
+        });
+    }
+}
