+++ TaskTok/templates/base.html
@@ -28,5 +28,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
             integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
             crossorigin="anonymous"></script>
+    <script type="module" src="{{ url_for('static', filename='js/utilities.js') }}"></script>
+   
 </body>
 </html>
