+++ TaskTok/templates/home.html
@@ -558,9 +558,7 @@ border-bottom: 0px solid #444444;
     <script src="https://cdn.jsdelivr.net/npm/flatpickr"></script>
 
     <!-- flatpickr used for date/time selection in date/offset input fields for add/edit forms -->
-    <script>
-
-</script>
+    <script type="module" src="{{ url_for('static', filename='js/apiRequests/test_functions.js') }}"></script>
 
 </body>
 {% endblock %}
